BACKEND/requirements.txt
alembic==1.14.1
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.43
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
